hackathon/MedGuideAI/requirements.txt
streamlit
python-dotenv
requests
Pillow
pinecone
openai
pymupdf